qualified post destroy method with 'if post' check in case Article was saved from other User

+++ app/controllers/articles_controller.rb
@@ -64,7 +64,11 @@ class ArticlesController < ApplicationController
     post = Post.where({email: user.email, a_url: article.a_url}).first
     # must also delete originating posts or update_articles will regenerate a SavedArticle
     # think about deleting posts the instance they are turned into an Article
-    post.destroy
+    
+    ##### TO DO #####
+    # need to change the update_articles method so that it doesnt pull from all posts
+    # this would allow us not to have to destroy posts
+    post.destroy if post
     saved_article.destroy
 
     respond_to do |format|
